Remove old logic
still issue with generate question button
it should only enable when any

do something with the button I am completely exhausted and lost in expla

Then why if I unchecked it stays still enabled but not working

This still nor works. Use a different approach

Instead of enabled disabled button use a simple warning to select the fi

Don't change the heading color to red it makes it difficult to read, als

Name of the chapters are still in red

Is the generate questions button work only once? As if I unchecked it no

Now I can't think of any reason for still not working. Give all possible

Remove the all logic for the generated questions button for now, we will

You didn't remove the error message from front end logic?

I see this error with the app, reported by NextJS, please fix it. The er

I see this error with the app, reported by NextJS, please fix it. The er

Still check uncheck in not working properly. It is perfectly working for

I see this error with the app, reported by NextJS, please fix it. The er

Change the id name to a short one

I see this error with the app, reported by NextJS, please fix it. The er

Now on generate questions button, check for that class and subject are s

As this message is shown then no need for the notification warning